Changed styles and UX, and added edit

+++ app.js
@@ -1,17 +1,33 @@
 function addItem(event){
     event.preventDefault();
-    const myList = document.getElementById('my-list');
-    const inputItem = document.getElementById('new-item');
-    const item = document.createElement('li');
-    const pItem = document.createElement('p')
-    const deleteItem = document.createElement('button')
-    deleteItem.innerHTML = 'X'
-    deleteItem.onclick = function () {
-        myList.removeChild(item)
+    const createInput = document.getElementById('new-item');
+    if(createInput.value !== ''){
+        const myList = document.getElementById('my-list');
+        const item = document.createElement('li');
+        const inputItem = document.createElement('input')
+        const deleteItem = document.createElement('button')
+        deleteItem.innerHTML = 'X'
+        deleteItem.onclick = function () {
+            myList.removeChild(item)
+        }
+        inputItem.value = createInput.value;
+        item.onclick = function () {
+            inputItem.disabled = false;
+            inputItem.focus();
+        }
+        inputItem.onblur = function () {
+            inputItem.disabled = true;
+        }
+        inputItem.onkeypress = function (event) {
+            if(event.keyCode == 13){
+                inputItem.disabled = true;
+            }
+        }
+        inputItem.disabled = true;
+        item.appendChild(inputItem)
+        item.appendChild(deleteItem)
+        myList.appendChild(item)
+        createInput.value = '';
     }
-    pItem.innerText = inputItem.value;
-    item.appendChild(pItem)
-    item.appendChild(deleteItem)
-    myList.appendChild(item)
-    inputItem.value = '';
+
 }
